Split branch and remote into modules
